backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
supabase==2.3.1
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
psycopg2-binary==2.9.9
email-validator==2.1.0
google-generativeai==0.3.2
python-docx==1.1.0
python-pptx==0.6.23
PyPDF2==3.0.1
openpyxl==3.1.2
python-magic==0.4.27
requests==2.31.0
aiofiles==23.2.1
